feat: add esm module
